requirements.txt
streamlit>=1.28.0
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.0
pydantic>=2.4.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
pydeck>=0.8.0
python-dotenv>=1.0.0
pytz>=2023.3
